+++ src/_actions/type.js
@@ -2,3 +2,4 @@ export const LOGIN_USER = "login_user"
 export const REGISTER_USER = "register_user"
 export const AUTH_USER = "auth_user"
 export const LOAD_MAP_DATA = "load_map_data"
+export const LOAD_ALARM = "load_alarm"
