--- src/redux/_actions/type.js
@@ -1,4 +1,0 @@
-export const LOGIN_USER = "login_user"
-export const REGISTER_USER = "register_user"
-export const AUTH_USER = "auth_user"
-export const LOAD_ALARM = "load_alarm"
